27.10.2022

+++ ft_strlen.c
@@ -1,6 +1,7 @@
-int ft_strlen(const char *s)
+#include "libft.h"
+size_t ft_strlen(const char *s)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while (s[i] != '\0')
